# docs/openapi.yaml
openapi: 3.1.0
info:
  title: 86.88 B&B — REST API
  version: 1.0.0
  description: |
    Backend API for the 86.88 民宿 (B&B) platform.

    Serves as the bridge between the PostgreSQL database and the frontend apps:
    - `apps/website` — public-facing Next.js site (8688bnb.com)
    - `apps/admin` — admin dashboard (admin.8688bnb.com)

    ## Authentication
    Admin authentication uses JWT tokens stored in `__Host-8688_session` HttpOnly cookies.
    All state-changing requests require a CSRF token via the `X-CSRF-Token` header.

    ## Rate Limiting
    - Auth endpoints: 20 requests / 15 minutes
    - General API: 200 requests / 15 minutes
    - Public reads: 500 requests / 15 minutes
    - Webhook ingress: 100 requests / 15 minutes

    Rate limit info is returned in response headers:
    - `X-RateLimit-Limit` — max requests in window
    - `X-RateLimit-Remaining` — requests left
    - `X-RateLimit-Reset` — window reset time (Unix epoch)

  contact:
    name: 8688 BnB Dev
    email: dev@8688bnb.com
  license:
    name: Private
    identifier: UNLICENSED
servers:
  - url: http://localhost:3333/api/v1
    description: Local development
  - url: https://api.8688bnb.com/api/v1
    description: Production (via Cloudflare Tunnel)
...
security: []
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: __Host-8688_session
      description: |
        JWT stored in a hardened HttpOnly cookie.
        Set automatically on successful login.
        Cookie attributes: HttpOnly, Secure, SameSite=Strict, Path=/
    csrfToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF token required for all state-changing requests (POST, PUT, DELETE).
        Obtained from the login response or `GET /auth/csrf-token`.
    webhookApiKey:
      type: apiKey
      in: header
      name: X-Webhook-Key
      description: |
        API key for authenticating incoming OTA channel manager webhooks.
        Configured in the WEBHOOK_API_KEY environment variable.

  # ─── Reusable Schemas ────────────────────────────────────────────────────────
  schemas:
...
    Room:
      type: object
      properties:
        id:
          type: integer
          example: 1
        slug:
          type: string
          pattern: '^[a-z][a-z0-9-]*$'
          example: r1
          description: URL-safe identifier (r1, r2, etc.)
        name_zh:
          type: string
          example: 浪漫情懷 景觀二人房
        name_en:
          type: string
          example: Scenic Double Room
        capacity:
          type: integer
          minimum: 1
          maximum: 10
          example: 2
        type:
          type: string
          enum: [double, quad, suite]
          example: double
        description:
          type: string
          example: 位於二樓的景觀房，可欣賞安農溪畔美景...
        amenities:
          type: array
          items:
            type: string
          example: [冷暖氣, 獨立衛浴, 免治馬桶, Wi-Fi, 冰箱]
        images:
          type: array
          items:
            $ref: '#/components/schemas/MediaItem'
          description: Room photo gallery
        price_weekday:
          type: number
          format: decimal
          example: 2600
          description: Weekday price (TWD)
        price_weekend:
          type: number
          format: decimal
          example: 3200
          description: Weekend price (TWD, Fri & Sat nights)
        price_holiday:
          type: number
          format: decimal
          example: 3600
          description: Holiday/peak season price (TWD)
        available:
          type: boolean
          example: true
          description: Whether this room is active/bookable
        sort_order:
          type: integer
          example: 1
          description: Display order on the website
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - slug
        - name_zh
        - capacity
        - type
        - price_weekday
        - price_weekend
        - price_holiday
        - available
        - sort_order
        - created_at
        - updated_at
    RoomCreateRequest:
      type: object
      properties:
        slug:
          type: string
          pattern: '^[a-z][a-z0-9-]*$'
          example: r6
        name_zh:
          type: string
          minLength: 1
          maxLength: 100
          example: 新房型名稱
        name_en:
          type: string
          maxLength: 100
          example: New Room Name
        capacity:
          type: integer
          minimum: 1
          maximum: 10
          example: 2
        type:
          type: string
          enum: [double, quad, suite]
          example: double
        description:
          type: string
          maxLength: 2000
        amenities:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 30
        price_weekday:
          type: number
          minimum: 0
          example: 2600
        price_weekend:
          type: number
          minimum: 0
          example: 3200
        price_holiday:
          type: number
          minimum: 0
          example: 3600
        available:
          type: boolean
          default: true
        sort_order:
          type: integer
          minimum: 0
          example: 6
      required: [slug, name_zh, capacity, type, price_weekday, price_weekend, price_holiday]
    RoomUpdateRequest:
      type: object
      properties:
        name_zh:
          type: string
          minLength: 1
          maxLength: 100
        name_en:
          type: string
          maxLength: 100
        capacity:
          type: integer
          minimum: 1
          maximum: 10
        type:
          type: string
          enum: [double, quad, suite]
        description:
          type: string
          maxLength: 2000
        amenities:
          type: array
          items:
            type: string
            maxLength: 50
          maxItems: 30
        price_weekday:
          type: number
          minimum: 0
        price_weekend:
          type: number
          minimum: 0
        price_holiday:
          type: number
          minimum: 0
        available:
          type: boolean
        sort_order:
          type: integer
          minimum: 0
      minProperties: 1
...
    MediaItem:
      type: object
      description: |
        A single image/media file.
        Media is organized by "target" — which page/section it belongs to.
        Target examples: homepage_hero, homepage_cats, homepage_bnb,
        about, rooms_overview, room_r1, room_r2, ..., location, gallery, etc.
      properties:
        id:
          type: integer
          example: 1
        target:
          type: string
          example: homepage_hero
          description: |
            Which page/section this media belongs to.
            Convention follows the site structure:
            - homepage_hero — Hero carousel images
            - homepage_8688 — 8688 section images
            - homepage_cats — Cats section images
            - homepage_bnb — BnB section images
            - about — About Us page
            - rooms_overview — Rooms overview (stay) page thumbnails
            - room_{slug} — Per-room gallery (e.g. room_r1, room_r2)
            - booking_info — Booking info / regulation page
            - location — Location page
            - gallery — General photo gallery
            - brand — Logo and brand assets
        filename_original:
          type: string
          example: room1-view.jpg
          description: Original uploaded filename (stored in DB only, not used for storage)
        filename_stored:
          type: string
          example: a3f8c2e1-4d5b-6789-0abc-def012345678.webp
          description: UUID-based filename on disk
        url:
          type: string
          format: uri
          example: /uploads/images/a3f8c2e1-4d5b-6789-0abc-def012345678.webp
          description: Public URL path to serve the image
        mime_type:
          type: string
          enum: [image/jpeg, image/png, image/webp]
          example: image/webp
        size_bytes:
          type: integer
          example: 245760
        width:
          type: integer
          nullable: true
          example: 1920
        height:
          type: integer
          nullable: true
          example: 1080
        alt_text:
          type: string
          nullable: true
          example: 浪漫情懷房間景觀
          description: Alt text for accessibility and SEO
        sort_order:
          type: integer
          example: 1
          description: Display order within the target group
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
        - id
        - target
        - filename_original
        - filename_stored
        - url
        - mime_type
        - size_bytes
        - sort_order
        - created_at
        - updated_at
...
    WebhookEvent:
      type: object
      description: |
        Logged event from an incoming OTA channel manager webhook.
        A single unified endpoint receives events from any channel manager
        (e.g. SiteMinder, Cloudbeds, or direct OTA integrations).
      properties:
        id:
          type: integer
          example: 1
        event_type:
          type: string
          enum: [new_booking, modify_booking, cancel_booking, rate_update, inventory_update, unknown]
          example: new_booking
        platform:
          type: string
          example: agoda
          description: Origin platform (extracted from payload or header)
        external_booking_id:
          type: string
          nullable: true
          example: AGD-123456
        payload:
          type: object
          additionalProperties: true
          description: Raw webhook payload stored for audit
        status:
          type: string
          enum: [received, processed, failed, ignored]
          example: received
        error_message:
          type: string
          nullable: true
          description: Error details if processing failed
        processed_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
      required: [id, event_type, platform, payload, status, created_at]
    WebhookIngestRequest:
      type: object
      description: |
        Unified webhook payload from a channel manager.
        The structure is flexible to accommodate different platforms.
      properties:
        event_type:
          type: string
          enum: [new_booking, modify_booking, cancel_booking, rate_update, inventory_update]
          example: new_booking
        platform:
          type: string
          maxLength: 50
          example: agoda
          description: Which OTA platform this event originated from
        external_booking_id:
          type: string
          maxLength: 100
          example: AGD-123456
        data:
          type: object
          additionalProperties: true
          description: |
            Platform-specific booking/event data.
            For new_booking: { guest_name, check_in, check_out, room_type, ... }
            For cancel_booking: { reason, ... }
            Stored as-is in the webhook log.
      required: [event_type, platform]

    # ── Dashboard ──────────────────────────────────────────────────────────────
...
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
          example: 1
        per_page:
          type: integer
          minimum: 1
          maximum: 100
          example: 20
        total:
          type: integer
          example: 47
        total_pages:
          type: integer
          example: 3
      required: [page, per_page, total, total_pages]
...
    ApiError:
      type: object
      description: Standard error response
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
              description: |
                Machine-readable error code. Examples:
                - AUTH_INVALID_CREDENTIALS
                - AUTH_SESSION_EXPIRED
                - AUTH_CSRF_INVALID
                - VALIDATION_ERROR
                - NOT_FOUND
                - RATE_LIMITED
                - CONFLICT (e.g. double-booking)
                - INTERNAL_ERROR
                - UPLOAD_FILE_TOO_LARGE
                - UPLOAD_INVALID_TYPE
            message:
              type: string
              example: 驗證失敗
              description: Human-readable error message (Chinese)
            message_en:
              type: string
              example: Validation failed
              description: English fallback message
            details:
              type: array
              items:
                type: object
                properties:
                  field:
                    type: string
                    example: check_in
                  message:
                    type: string
                    example: 入住日期不得早於今天
              description: Per-field validation errors (if applicable)
          required: [code, message]
      required: [success, error]
...
  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number for pagination
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    RoomSlugParam:
      name: slug
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z][a-z0-9-]*$'
      description: Room slug (e.g. r1, r2)
...
    DateFromParam:
      name: from
      in: query
      required: true
      schema:
        type: string
        format: date
      description: Start date (inclusive)
    DateToParam:
      name: to
      in: query
      required: true
      schema:
        type: string
        format: date
      description: End date (inclusive)

  # ─── Reusable Responses ──────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Authentication required or session expired
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: AUTH_SESSION_EXPIRED
              message: 登入已過期，請重新登入
              message_en: Session expired, please log in again
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: 找不到資源
              message_en: Resource not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: 資料驗證失敗
              message_en: Validation failed
              details:
                - field: check_in
                  message: 入住日期不得早於今天
    RateLimited:
      description: Too many requests
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiError'
          example:
            success: false
            error:
              code: RATE_LIMITED
              message: 請求太頻繁，請稍後再試
              message_en: Too many requests, please try again later
      headers:
        Retry-After:
          schema:
            type: integer
          description: Seconds to wait before retrying
...
paths:
...
  /rooms:
    get:
      tags: [Rooms]
      operationId: listRooms
      summary: List all rooms
      description: |
        Public endpoint — returns all active rooms for the website.
        Admin users see all rooms including unavailable ones.
      parameters:
        - name: include_unavailable
          in: query
          schema:
            type: boolean
            default: false
          description: Include unavailable rooms (admin only)
      responses:
        '200':
          description: Room list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Room'
                required: [success, data]

    post:
      tags: [Rooms]
      operationId: createRoom
      summary: Create a new room
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomCreateRequest'
      responses:
        '201':
          description: Room created
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Room'
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: Room slug already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: CONFLICT
                  message: 此房型代碼已存在
                  message_en: Room slug already exists
  /rooms/{slug}:
    parameters:
      - $ref: '#/components/parameters/RoomSlugParam'

    get:
      tags: [Rooms]
      operationId: getRoom
      summary: Get room detail
      description: Public endpoint — returns room details by slug.
      responses:
        '200':
          description: Room detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Room'
                required: [success, data]
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Rooms]
      operationId: updateRoom
      summary: Update room
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoomUpdateRequest'
      responses:
        '200':
          description: Room updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Room'
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Rooms]
      operationId: deleteRoom
      summary: Soft-delete room
      description: |
        Marks the room as unavailable (soft delete).
        Existing bookings are not affected.
      security:
        - cookieAuth: []
          csrfToken: []
      responses:
        '200':
          description: Room soft-deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      message:
                        type: string
                        example: 房型已停用
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /rooms/{slug}/availability:
    parameters:
      - $ref: '#/components/parameters/RoomSlugParam'
    get:
      tags: [Rooms]
      operationId: checkRoomAvailability
      summary: Check room availability
      description: |
        Public endpoint — checks if a room is available for a date range.
        Returns availability status considering both bookings and blocked dates.
      parameters:
        - $ref: '#/components/parameters/DateFromParam'
        - $ref: '#/components/parameters/DateToParam'
      responses:
        '200':
          description: Availability result
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      available:
                        type: boolean
                        example: true
                      room_slug:
                        type: string
                        example: r1
                      from:
                        type: string
                        format: date
                      to:
                        type: string
                        format: date
                      conflicts:
                        type: array
                        items:
                          type: object
                          properties:
                            date:
                              type: string
                              format: date
                            reason:
                              type: string
                              enum: [booked, blocked]
                            message:
                              type: string
                              description: If blocked, shows the reason/comment
                        description: Dates that are unavailable (if any)
                      estimated_price:
                        type: number
                        format: decimal
                        description: Estimated total price for the date range (TWD)
                    required: [available, room_slug, from, to]
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # BOOKINGS
  # ═══════════════════════════════════════════════════════════════════════════
...
  /webhooks/ingest:
    post:
      tags: [Webhooks]
      operationId: ingestWebhook
      summary: Receive OTA channel manager webhook
      description: |
        Unified webhook endpoint for receiving booking events from
        channel managers (e.g. SiteMinder, Cloudbeds) or direct OTA
        integrations (Agoda, Booking.com, Trip.com, etc.).

        A single endpoint handles all platforms — the `platform` field
        in the payload identifies the source.

        Events are logged and can be auto-processed or manually reviewed
        by the admin.

        **Authentication:** Requires the `X-Webhook-Key` header matching
        the configured WEBHOOK_API_KEY.
      security:
        - webhookApiKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookIngestRequest'
      responses:
        '200':
          description: Webhook received and logged
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      event_id:
                        type: integer
                        description: ID of the created webhook event log entry
                      status:
                        type: string
                        enum: [received, processed]
                    required: [event_id, status]
                required: [success, data]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid or missing API key
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
              example:
                success: false
                error:
                  code: AUTH_INVALID_CREDENTIALS
                  message: Webhook API key 無效
                  message_en: Invalid webhook API key
        '429':
          $ref: '#/components/responses/RateLimited'
  /webhooks/events:
    get:
      tags: [Webhooks]
      operationId: listWebhookEvents
      summary: List webhook event log
      description: |
        View all received webhook events with filtering.
        Admin only — used for debugging and audit.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - name: status
          in: query
          schema:
            type: string
            enum: [received, processed, failed, ignored]
        - name: platform
          in: query
          schema:
            type: string
          description: Filter by source platform
        - name: event_type
          in: query
          schema:
            type: string
            enum: [new_booking, modify_booking, cancel_booking, rate_update, inventory_update, unknown]
      responses:
        '200':
          description: Webhook event list
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/WebhookEvent'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
                required: [success, data, meta]
        '401':
          $ref: '#/components/responses/Unauthorized'
  /webhooks/events/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          minimum: 1
        description: Webhook event ID

    get:
      tags: [Webhooks]
      operationId: getWebhookEvent
      summary: Get webhook event detail
      description: View full payload and processing status of a webhook event.
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Webhook event detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/WebhookEvent'
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Webhooks]
      operationId: updateWebhookEventStatus
      summary: Update webhook event status
      description: |
        Manually mark a webhook event as processed, ignored, etc.
        Used by admin to manage unprocessed events.
      security:
        - cookieAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [processed, ignored, failed]
                error_message:
                  type: string
                  maxLength: 500
              required: [status]
      responses:
        '200':
          description: Event status updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/WebhookEvent'
                required: [success, data]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ═══════════════════════════════════════════════════════════════════════════
  # SYSTEM
  # ═══════════════════════════════════════════════════════════════════════════
...